Added xterm support and defaults
Now that I am using XMonad as my windowing system and I am trying to rid
myself of using a mouse, I have switched to xterm over the default gnome
terminal emulator. This helps remove the chrome and what have you. As such,
I have added a .Xresources file to determine the look of the terminal and
added it as the default in my XMonad file to launch with the default key
bindings.