Updated .bashrc and added .pryrc
I was including /etc/bash.bashrc in my .bashrc file. But, this was not good
since this file does not exist on the Mac (or other systems I suppose) and
as far as I know, my debian box is the only one with that file as of now.
Anyways, I added a check to see if it exists and is a file before sourcing
it.

I addd a .pryrc file. Really, this is the same as the .irbrc except that I
will be using pry instead of irb. It's just a lot nicer to work with.