Renamed 'extra' bash load file
There was an extra file that I was checking for and loading
in my .bashrc to load system-specific configurations that
don't really belong in my shared .bashrc file. However, I was
including .bash_profile as my extra. This file was the one
including my .bashrc on some of my machines. This caused it to
halt on boot (cyclic loading) so I have renamed this _extra_
bash file to `.bash_etc`